+++ src/builders/mod.rs
@@ -1,7 +1,9 @@
 /// This module contains all supported builder
 /// implementations for Liveboat - these contain functionality
-/// used for processing and outputting page files and 
+/// used for processing and outputting page files and
 /// rendering templates.
 
 pub mod spa_builder;
 pub mod aux;
+
+mod utils;