+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Refactored command handlers into separate module
 - Added builders module
- 
+- Update colors for default template
+
 ## [1.0.3] 2024-12-23
 ### Changed
 - Use crates.io version of libnewsboat